Move supabase calls to supabaseClient
